Pull Allocations out of lower loops into its own pass (#587)
* Move allocation into its own pass, allocate closer to where required.

* Move expression sorting out of lower loops. (#588)
