+++ torch/csrc/jit/codegen/cuda/lower_loops.h
@@ -41,14 +41,6 @@ class TORCH_CUDA_API LoopNestGenerator {
  private:
   LoopNestGenerator(Fusion* fusion, const std::vector<Expr*>& exprs);
 
-  // Create the allocation for tv, place it inside the loop associated with
-  // alloc_id, return the node
-  kir::Expr* pushAlloc(TensorView*);
-
-  // Track dynamic shared memory buffers
-  // Insert allocation at the beginning of the kernel
-  std::deque<kir::Allocate*> dynamic_smem_;
-
   // Open a new inner most for loop, track which TV it was constructed from
   // according to the computeAt chain.
   void openFor(IterDomain*);
@@ -59,24 +51,12 @@ class TORCH_CUDA_API LoopNestGenerator {
   // Appends an expression to the current scope
   void pushBack(kir::Expr* expr);
 
-  // Initialize a buffer to init_val. If this buffer is in smem or registers,
-  // pass in its allocation statement so we can make sure that we insert this
-  // initialization after the allocation.
-  void initReduction(TensorView* tv, Val* init_val, kir::Expr* alloc_expr);
-
   void handle(const Expr*);
 
   // Run the pass and accumulate output in lowered_exprs_
   void generate(const std::vector<Expr*>& exprs);
 
  private:
-  // Track number of allocations in each for loop. It is used to insert
-  // allocations in the correct order, which is necessary for memory aliasing
-  std::unordered_map<kir::ForLoop*, size_t> for_loop_allocations_;
-
-  // Track number of allocations outside any for loop.
-  size_t lowered_exprs_allocations_ = 0;
-
   // Lowered exprs to return
   std::vector<kir::Expr*> lowered_exprs_;
